pyproject.toml
[tool.poetry]
name = "gradient-color-palette"
version = "0.1.0"
description = ""
authors = ["b-basa"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pillow = "^11.1.0"
numpy = "^2.2.3"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119